Cleaned up buy machine guard extraction
