+++ src/state/machines/buy.ts
@@ -1,35 +1,8 @@
 import { AnimalCard, Card, ElementCard, GameState, PlantCard } from "@/state/types";
-import { compact, countBy, every, find, intersection, without } from "lodash";
+import { find, without } from "lodash";
 import { setup, assign, sendParent } from "xstate";
 import { produce } from "immer";
-
-export const BuyMachineGuards = {
-  canBuyCard: ({ context: { players, turn } }: { context: GameState }, card: AnimalCard | PlantCard) => {
-    if (card.type === "plant") {
-      const elementsCounted = countBy(card.elements);
-      const player = find(players, { uid: turn.player })!;
-      const playedElements =
-        compact([...player.hand, turn.borrowedElement])
-          .filter(({ uid }) => turn.playedCards.includes(uid) || turn.borrowedElement?.uid === uid)
-          .filter(({ type }) => type === "element") ?? [];
-      const playedElementsCounted = countBy(playedElements, "name");
-
-      return every(elementsCounted, (count, element) => playedElementsCounted[element] >= count);
-    }
-
-    if (card.type === "animal") {
-      const requiredBiomes = card.biomes;
-      const playedPlants =
-        (find(players, { uid: turn.player })
-          ?.hand.filter(({ uid }) => turn.playedCards.includes(uid))
-          .filter(({ type }) => type === "plant") as PlantCard[]) ?? [];
-
-      return playedPlants.filter(({ biomes }) => intersection(biomes, requiredBiomes).length > 0).length >= 2;
-    }
-
-    return false;
-  },
-};
+import { BuyMachineGuards } from "./guards/buy";
 
 export type BuyMachineOutEvents =
   | { type: "buy.card"; card: PlantCard | AnimalCard }
@@ -70,23 +43,9 @@ export const BuyMachine = setup({
     ),
   },
   guards: {
-    canBorrow: ({ context: { turn } }: { context: GameState }) => turn.borrowedCount < turn.borrowedLimit,
-    notPlayedCard: (
-      {
-        context: {
-          turn: { playedCards },
-        },
-      }: { context: GameState },
-      uid: Card["uid"],
-    ) => !playedCards.includes(uid),
-    notExhausted: (
-      {
-        context: {
-          turn: { exhaustedCards },
-        },
-      }: { context: GameState },
-      uid: Card["uid"],
-    ) => !exhaustedCards.includes(uid),
+    notPlayedCard: BuyMachineGuards.notPlayedCard,
+    notExhausted: BuyMachineGuards.notExhausted,
+    canBorrow: BuyMachineGuards.canBorrow,
     canBuyCard: BuyMachineGuards.canBuyCard,
   },
 }).createMachine({
